+++ app/Filament/Resources/ClientPositionResource.php
@@ -2,28 +2,27 @@
 
 namespace App\Filament\Resources;
 
+use App\Concerns\Filament\RequiresSuperAdminForClientMenu;
+use App\Enums\CRoleAssignation;
 use App\Filament\Resources\ClientPositionResource\Pages;
-use App\Filament\Resources\ClientPositionResource\RelationManagers;
 use App\Models\ClientPosition;
+use App\Models\CRole;
+use App\Models\CRoleLevel;
 use Filament\Forms;
 use Filament\Forms\Form;
 use Filament\Resources\Resource;
 use Filament\Tables;
 use Filament\Tables\Table;
+use Hugomyb\FilamentMediaAction\Tables\Actions\MediaAction;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\SoftDeletingScope;
-use App\Enums\CRoleAssignation;
-use App\Models\CRoleLevel;
-use App\Models\CRole;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Storage;
-use Hugomyb\FilamentMediaAction\Tables\Actions\MediaAction;
-
 
 class ClientPositionResource extends Resource
 {
-    protected static ?string $model = ClientPosition::class;
+    use RequiresSuperAdminForClientMenu;
 
+    protected static ?string $model = ClientPosition::class;
 
     protected static ?string $navigationLabel = 'Riwayat Jabatan';
 
@@ -60,10 +59,9 @@ class ClientPositionResource extends Resource
 
                 Forms\Components\Select::make('c_role_level_id')
                     ->label(__('labels.form.client.fields.crole_grade'))
-                    ->options(fn (Forms\Get $get) =>
-                        CRoleLevel::query()
-                            ->where('c_role_id', $get('c_role_id'))
-                            ->pluck('level', 'id')
+                    ->options(fn (Forms\Get $get) => CRoleLevel::query()
+                        ->where('c_role_id', $get('c_role_id'))
+                        ->pluck('level', 'id')
                     )
                     ->disabled(fn (Forms\Get $get) => blank($get('c_role_id')))
                     ->required(),
@@ -73,25 +71,24 @@ class ClientPositionResource extends Resource
                     ->label(__('labels.form.client.fields.assignation_type'))
                     ->required(),
                 Forms\Components\DatePicker::make('effective_date')
-                     ->label('TMT Jabatan')
+                    ->label('TMT Jabatan')
                     ->required(),
                 Forms\Components\TextInput::make('decree_number')
                     ->label('Nomor SK Jabatan')
                     ->required()
                     ->maxLength(255),
                 Forms\Components\FileUpload::make('decree_file')
-                            ->disk('s3')
-                            ->label('File SK Jabatan')
-                            ->required()
-                            ->maxFiles(1)
-                            ->acceptedFileTypes(config('fungsional-pro.accepted_document_type'))
-                            ->maxSize(config('fungsional-pro.max_upload_file_size'))
-                            ->directory('decree_file')
-                            ->visibility('private')
-                            ->downloadable(),
+                    ->disk('s3')
+                    ->label('File SK Jabatan')
+                    ->required()
+                    ->maxFiles(1)
+                    ->acceptedFileTypes(config('fungsional-pro.accepted_document_type'))
+                    ->maxSize(config('fungsional-pro.max_upload_file_size'))
+                    ->directory('decree_file')
+                    ->visibility('private')
+                    ->downloadable(),
             ]);
     }
-
 
     public static function table(Table $table): Table
     {
@@ -134,7 +131,7 @@ class ClientPositionResource extends Resource
             ])
             ->actions([
                 MediaAction::make()
-                    ->media(fn(Model $record) => Storage::temporaryUrl($record->decree_file, now()->addMinutes(10)))
+                    ->media(fn (Model $record) => Storage::temporaryUrl($record->decree_file, now()->addMinutes(10)))
                     ->label('SK Jabatan'),
                 Tables\Actions\EditAction::make(),
             ])
@@ -166,5 +163,4 @@ class ClientPositionResource extends Resource
     {
         return __('labels.nav.client_menu');
     }
-
 }