+++ app/Filament/Resources/ClientPositionResource/Pages/EditClientPosition.php
@@ -2,12 +2,15 @@
 
 namespace App\Filament\Resources\ClientPositionResource\Pages;
 
+use App\Concerns\Filament\RedirectsLockedClientMenuAccess;
 use App\Filament\Resources\ClientPositionResource;
 use Filament\Actions;
 use Filament\Resources\Pages\EditRecord;
 
 class EditClientPosition extends EditRecord
 {
+    use RedirectsLockedClientMenuAccess;
+
     protected static string $resource = ClientPositionResource::class;
 
     protected function getHeaderActions(): array